chore(tests): Use more recent syntax (01_basic.rs)
