--- tests/should_succeed/cell/01/proof.json
@@ -1,15 +1,0 @@
-{
-  "profile": [
-    { "prover": "z3@4.12.4", "size": 33, "time": 0.461 },
-    { "prover": "cvc5@1.0.5", "size": 46, "time": 0.578 },
-    { "prover": "alt-ergo@2.6.0", "size": 17, "time": 0.512 },
-    { "prover": "cvc4@1.8", "size": 45, "time": 0.39 }
-  ],
-  "proofs": {
-    "M_01__adds_two": {
-      "vc_adds_two": { "prover": "cvc5@1.0.5", "time": 0.03 },
-      "vc_get": { "prover": "cvc5@1.0.5", "time": 0.014 },
-      "vc_set": { "prover": "cvc5@1.0.5", "time": 0.033 }
-    }
-  }
-}
