+++ tests/should_succeed/cell/01_basic/proof.json
@@ -1,0 +1,15 @@
+{
+  "profile": [
+    { "prover": "z3@4.12.4", "size": 33, "time": 0.623 },
+    { "prover": "cvc5@1.0.5", "size": 46, "time": 0.533 },
+    { "prover": "alt-ergo@2.6.0", "size": 17, "time": 0.61 },
+    { "prover": "cvc4@1.8", "size": 45, "time": 0.426 }
+  ],
+  "proofs": {
+    "M_01_basic__adds_two": {
+      "vc_adds_two": { "prover": "cvc5@1.0.5", "time": 0.029 },
+      "vc_get": { "prover": "cvc5@1.0.5", "time": 0.026 },
+      "vc_set": { "prover": "cvc5@1.0.5", "time": 0.02 }
+    }
+  }
+}
